+++ CHANGELOG.md
@@ -4,6 +4,23 @@
 
 ### Features
 
+- Add screenshot masking support using view hierarchy ([#5073](https://github.com/getsentry/sentry-java/pull/5073))
+  - Masks sensitive content (text, images) in error screenshots before sending to Sentry
+  - Reuses Session Replay's masking logic; requires `sentry-android-replay` module at runtime
+  - To enable masking programmatically:
+    ```kotlin
+    SentryAndroid.init(context) { options ->
+        options.isAttachScreenshot = true
+        options.screenshotOptions.setMaskAllText(true)
+        options.screenshotOptions.setMaskAllImages(true)
+    }
+    ```
+  - Or via AndroidManifest.xml:
+    ```xml
+    <meta-data android:name="io.sentry.attach-screenshot" android:value="true" />
+    <meta-data android:name="io.sentry.screenshot.mask-all-text" android:value="true" />
+    <meta-data android:name="io.sentry.screenshot.mask-all-images" android:value="true" />
+    ```
 - Add `installGroupsOverride` parameter and `installGroups` property to Build Distribution SDK ([#5062](https://github.com/getsentry/sentry-java/pull/5062))
 - Update Android targetSdk to API 36 (Android 16) ([#5016](https://github.com/getsentry/sentry-java/pull/5016))
 - Add AndroidManifest support for Spotlight configuration via `io.sentry.spotlight.enable` and `io.sentry.spotlight.url` ([#5064](https://github.com/getsentry/sentry-java/pull/5064))